Document 1 more method + fix delete behaviour

+++ Backend/DbWrapper/DbWrapper.py
@@ -73,6 +73,9 @@ class DbWrapper:
         self.db.collection(COURSES).document(course_id).set(template)
         return True
     def removeCourse(self, course_id:str)->bool:
+        x = [i for i in self.db.collection(COURSES).where(filter=FieldFilter("course_id", "==", course_id)).stream()]
+        if len(x) == 0:
+            return False
         try:
             db.collection(COURSES).document(course_id).delete()
         except:
